+++ lib/cli/Cargo.toml
@@ -24,12 +24,19 @@ doc = false
 required-features = ["headless"]
 
 [dependencies]
-wasmer-artifact = { version = "=2.2.1", path = "../artifact" }
+wasmer = { version = "=2.2.1", path = "../api", default-features = false }
 wasmer-compiler = { version = "=2.2.1", path = "../compiler" }
+wasmer-compiler-cranelift = { version = "=2.2.1", path = "../compiler-cranelift", optional = true }
 wasmer-compiler-singlepass = { version = "=2.2.1", path = "../compiler-singlepass", optional = true }
+wasmer-compiler-llvm = { version = "=2.2.1", path = "../compiler-llvm", optional = true }
+wasmer-emscripten = { version = "=2.2.1", path = "../emscripten", optional = true }
 wasmer-engine = { version = "=2.2.1", path = "../engine" }
 wasmer-engine-universal = { version = "=2.2.1", path = "../engine-universal", optional = true }
+wasmer-engine-dylib = { version = "=2.2.1", path = "../engine-dylib", optional = true }
+wasmer-engine-staticlib = { version = "=2.2.1", path = "../engine-staticlib", optional = true }
+wasmer-vm = { version = "=2.2.1", path = "../vm" }
 wasmer-wasi = { version = "=2.2.1", path = "../wasi", optional = true }
+wasmer-wasi-experimental-io-devices = { version = "=2.2.1", path = "../wasi-experimental-io-devices", optional = true }
 wasmer-wast = { version = "=2.2.1", path = "../../tests/lib/wast", optional = true }
 wasmer-cache = { version = "=2.2.1", path = "../cache", optional = true }
 wasmer-types = { version = "=2.2.1", path = "../types" }
@@ -47,19 +54,6 @@ cfg-if = "1.0"
 fern = { version = "0.6", features = ["colored"], optional = true }
 log = { version = "0.4", optional = true }
 tempfile = "3"
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-wasmer = { version = "=2.2.1", path = "../api", default-features = false, features = ["js-default"] }
-
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-wasmer = { version = "=2.2.1", path = "../api", default-features = false }
-wasmer-vm = { version = "=2.2.1", path = "../vm" }
-wasmer-compiler-cranelift = { version = "=2.2.1", path = "../compiler-cranelift", optional = true }
-wasmer-compiler-llvm = { version = "=2.2.1", path = "../compiler-llvm", optional = true }
-wasmer-wasi-experimental-io-devices = { version = "=2.2.1", path = "../wasi-experimental-io-devices", optional = true }
-wasmer-emscripten = { version = "=2.2.1", path = "../emscripten", optional = true }
-wasmer-engine-dylib = { version = "=2.2.1", path = "../engine-dylib", optional = true }
-wasmer-engine-staticlib = { version = "=2.2.1", path = "../engine-staticlib", optional = true }
 
 [target.'cfg(target_os = "linux")'.dependencies]
 unix_mode = "0.1.3"
