+++ lib/types/src/lib.rs
@@ -100,9 +100,9 @@ pub use types::{
 pub use archives::ArchivableIndexMap;
 
 pub use crate::libcalls::LibCall;
-pub use crate::memory::{MemoryError, MemoryStyle};
+pub use crate::memory::MemoryStyle;
 pub use crate::table::TableStyle;
-pub use crate::trapcode::{Trap, TrapCode};
+pub use crate::trapcode::TrapCode;
 pub use crate::vmoffsets::{TargetSharedSignatureIndex, VMBuiltinFunctionIndex, VMOffsets};
 
 pub use crate::utils::is_wasm;
