+++ lib/vm/Cargo.toml
@@ -26,8 +26,6 @@ loupe = { version = "0.1", features = ["enable-indexmap"] }
 enum-iterator = "0.7.0"
 scopeguard = "1.1.0"
 lazy_static = "1.4.0"
-
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 region = { version = "3.0" }
 corosensei = { version = "0.1.2" }
