+++ lib/vm/src/lib.rs
@@ -43,7 +43,7 @@ pub use crate::instance::{
     ImportFunctionEnv, ImportInitializerFuncPtr, InstanceAllocator, InstanceHandle,
     WeakOrStrongInstanceRef,
 };
-pub use crate::memory::{LinearMemory, Memory};
+pub use crate::memory::{LinearMemory, Memory, MemoryError};
 pub use crate::mmap::Mmap;
 pub use crate::probestack::PROBESTACK;
 pub use crate::sig_registry::SignatureRegistry;
@@ -56,13 +56,13 @@ pub use crate::vmcontext::{
 };
 pub use wasmer_artifact::{FunctionBodyPtr, VMFunctionBody};
 pub use wasmer_types::LibCall;
+pub use wasmer_types::MemoryStyle;
 pub use wasmer_types::TableStyle;
 #[deprecated(
     since = "2.1.0",
     note = "ModuleInfo, ExportsIterator, ImportsIterator should be imported from wasmer_types."
 )]
 pub use wasmer_types::{ExportsIterator, ImportsIterator, ModuleInfo};
-pub use wasmer_types::{MemoryError, MemoryStyle};
 pub use wasmer_types::{
     TargetSharedSignatureIndex, VMBuiltinFunctionIndex, VMExternRef, VMOffsets,
 };
